+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 ### Added
 
-- Add handling for `-T` TCP.IP switch on EventStore args [#46](https://github.com/jet/dotnet-templates/pulls/46)
+- Add handling for `-T` TCP/IP switch on EventStore args [#46](https://github.com/jet/dotnet-templates/pulls/46)
 
 ### Changed
 ### Removed