+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The `Unreleased` section name is replaced by the expected version of next releas
 - Add handling for `-T` TCP/IP switch on EventStore args [#46](https://github.com/jet/dotnet-templates/pulls/46)
 
 ### Changed
+
+- Target `Propulsion`.*, `Equinox`.*, `FsCodec`.* v `2.0.0`
+
 ### Removed
 ### Fixed